@@ -2,7 +2,7 @@ package com.dfl.openipma.base
 
 import android.annotation.SuppressLint
 import android.content.Intent
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.view.MenuItem
 import com.dfl.openipma.R
 import com.dfl.openipma.about.AboutActivity
