@@ -1,14 +1,14 @@
 package com.dfl.openipma.base
 
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Dispatchers.Main
 import kotlinx.coroutines.Job
 
 open class BaseViewModel : ViewModel() {
 
     private val viewModelJob = Job()
-    val scope = CoroutineScope(Dispatchers.Main + viewModelJob)
+    val scope = CoroutineScope(Main + viewModelJob)
 
     override fun onCleared() {
         super.onCleared()
